+++ CHANGELOG.md
@@ -133,6 +133,9 @@ format.
 - Rewrite SVG `feImage` asset references discovered in exported pages and
   linked copied SVG files.
   ([#3619](https://github.com/WordPress/wordpress-playground/pull/3619))
+- Rewrite Windows browser config metadata and copied XML tile asset
+  dependencies in static exports.
+  ([#3619](https://github.com/WordPress/wordpress-playground/pull/3619))
 
 ### Documentation